pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "python-multi-tenant-auth-template"
version = "0.1.0"
description = "FastAPI multi-tenant authentication template"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=3.0.2,<4.0.0",
    "alembic>=1.13.2,<2.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "bcrypt>=4.1.3,<5.0.0",
    "email-validator>=2.2.0,<3.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "pydantic-settings>=2.4.0,<3.0.0",
    "PyJWT>=2.9.0,<3.0.0",
    "SQLAlchemy>=2.0.32,<3.0.0",
    "structlog>=24.4.0,<26.0.0",
    "uvicorn[standard]>=0.30.5,<1.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
